20260128_정우철_bom 자동 등록

+++ src/main/java/mes/domain/repository/BomComponentRepositoryCustom.java
@@ -3,9 +3,10 @@ package mes.domain.repository;
 import mes.domain.entity.BomComponent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 public interface BomComponentRepositoryCustom {
 
-    void insertIfNotExists(Long bomId, Integer materialId, int amount);
+    void insertIfNotExists(Long bomId, Integer materialId, BigDecimal amount);
 }
